fix: fix CI build issues
This PR attempts to fix the CI build issues stemming from a number of old dependencies that are now conflicting.

Update the package.json file to match versions of core dependencies used in another MFE (using frontend-app-communications as a starting point). This also removes the use of the edx.org styling for the header and footer by default. We are already injecting/overwriting the edx.org styled header and footer during deployment and won't cause styling issues in stage or prod.