feat!: Remove unused feature flag
[APER-2922]

This PR removes the `USE_LR_MFE` feature flag. This feature flag was initially added to redirect learners back to the legacy frontend in Credentials when viewing a Program Record. This was so we could release the Learner Record MFE in phases while we were implementing the design.

Now that the legacy frontend has been removed from Credentials, this functionality and feature flag is no longer needed.